+++ sdks/python/apache_beam/examples/snippets/snippets.py
@@ -31,15 +31,9 @@ string. The tags can contain only letters, digits and _.
 """
 # pytype: skip-file
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import argparse
 import base64
 import json
-from builtins import object
-from builtins import range
 from decimal import Decimal
 
 import apache_beam as beam