+++ sdks/python/apache_beam/internal/util_test.py
@@ -18,8 +18,6 @@
 """Unit tests for the util module."""
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import unittest
 
 from apache_beam.internal.util import ArgumentPlaceholder